Исправлена сборка под Windows. Merge branch '2026-07/build_fix_001'.

+++ common/trunklink.cpp
@@ -210,7 +210,7 @@ void TrunkLink::SendPacketQueue() {
     auto info = *it;
 
     auto ava = GetAvailableBuffer(info.CtxID);
-    if (ava < info.PacketSize) {
+    if (ava < static_cast<int>(info.PacketSize)) {
       ++it;
       continue;
     }
@@ -731,9 +731,9 @@ int TrunkClient::GetAvailableBuffer(ConnectID ctx) {
       curt - trunk_buffer_last_time_)
                   .count();
 
-  trunk_buffer_last_size_ +=
-      intr * kDefaultUdpTrafficSpeed;  // Учитываем, что со временем буфер
-                                       // освобождается
+  trunk_buffer_last_size_ += static_cast<int>(
+      intr * kDefaultUdpTrafficSpeed);  // Учитываем, что со временем буфер
+                                        // освобождается
   if (trunk_buffer_last_size_ > trunk_socket_buffer_size_) {
     trunk_buffer_last_size_ = trunk_socket_buffer_size_;
   }
@@ -902,9 +902,9 @@ int TrunkServer::GetAvailableBuffer(ConnectID ctx) {
       curt - ts.buffer_last_time_)
                   .count();
 
-  ts.buffer_last_size_ +=
-      intr * kDefaultUdpTrafficSpeed;  // Учитываем, что со временем буфер
-                                       // освобождается
+  ts.buffer_last_size_ += static_cast<int>(
+      intr * kDefaultUdpTrafficSpeed);  // Учитываем, что со временем буфер
+                                        // освобождается
   if (ts.buffer_last_size_ > ts.socket_buffer_size_) {
     ts.buffer_last_size_ = ts.socket_buffer_size_;
   }
